requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_bcrypt
pyjwt
reportlab
